Remove DatumWithScore, deprecated BatchTrainScore, move DatumWithNorm inside OutlierClassifier
